requirements.txt
matplotlib
pillow
numpy
pandas
requests
